Rely more on rationals
Fix a couple of broken tests.
