+++ tests/multiply.test.ts
@@ -19,7 +19,7 @@ let examples = [
     [
         "123456789.987654321",
         "987654321.123456789",
-        "121932632103337905.6620941931126352",
+        "121932632103337905.6620941931126353",
     ],
 ];
